begin refactor to use constants file
