+++ components/constants.ts
@@ -1,0 +1,15 @@
+const constants = {
+    me: {
+        "fname": "Aaron",
+        "lname": "Mechanic",
+    },
+    school: {},
+    currentJob: {
+        "positionTitle": "",
+        "companyName": "",
+    },
+
+
+}
+
+export default constants;
