--- docs/_next/static/4K2Yjteo4Dxw2SGeeA2zo/_buildManifest.js
@@ -1,1 +1,0 @@
-self.__BUILD_MANIFEST=function(s){return{__rewrites:{beforeFiles:[],afterFiles:[],fallback:[]},"/":["static/chunks/a9a7754c-6e4c9bcf2e5d0488.js",s,"static/css/27521a07fd2232a3.css","static/chunks/pages/index-7ae8e968a397b372.js"],"/_error":["static/chunks/pages/_error-1995526792b513b2.js"],"/projects":[s,"static/chunks/12-5f8a961c82e975b7.js","static/css/7e39b8563f411134.css","static/chunks/pages/projects-3bffc8a2dcdae0fe.js"],"/resume":["static/chunks/547-6827b4548ab51a48.js","static/css/ab5bfa68abd02f74.css","static/chunks/pages/resume-50ac787ef8fcfbcf.js"],sortedPages:["/","/_app","/_error","/projects","/resume"]}}("static/chunks/374-9c4b8e10ece95bb1.js"),self.__BUILD_MANIFEST_CB&&self.__BUILD_MANIFEST_CB();
