--- docs/_next/static/chunks/pages/index-7ae8e968a397b372.js
@@ -1,1 +1,0 @@
-(self.webpackChunk_N_E=self.webpackChunk_N_E||[]).push([[405],{5301:function(e,t,r){(window.__NEXT_P=window.__NEXT_P||[]).push(["/",function(){return r(3317)}])},3317:function(e,t,r){"use strict";r.r(t),r.d(t,{default:function(){return J}});var a=r(5893),n=r(3165),o=r(9008),i=r(7294),s=r(275);var l=(0,s.zo)("span",{dflex:"center",position:"relative",zIndex:"$1",boxSizing:"border-box",overflow:"hidden",verticalAlign:"top",cursor:"auto",".nextui-avatar-bg":{position:"absolute",top:0,left:0,right:0,bottom:0,zIndex:"$2",transition:"$avatar",size:"100%"},"&:hover .nextui-avatar-bg":{boxShadow:" inset 0 0 40px 0 rgb(0 0 0 / 14%)"},".nextui-avatar-img":{opacity:0,zIndex:"$3",display:"flex",bg:"$background",transition:"transform 250ms ease 0ms, opacity 200ms ease-in 0ms",width:"100%",height:"100%",objectFit:"cover"},'&[data-state="ready"] .nextui-avatar-img':{opacity:1},".nextui-avatar-icon":{display:"flex",position:"absolute",left:"50%",top:"50%",ta:"center",zIndex:"$2",transform:"translate(-50%, -50%)",whiteSpace:"nowrap",us:"none"},".nextui-avatar-text":{position:"absolute",zIndex:"$2",left:"50%",top:"50%",ta:"center",color:"$text",fontWeight:"$medium",transform:"translate(-50%, -50%) scale(0.65)",whiteSpace:"nowrap",us:"none"},"@motion":{".nextui-avatar-bg, .nextui-avatar-img":{transition:"none"}},variants:{color:{default:{".nextui-avatar-bg":{bg:"$accents2"}},primary:{".nextui-avatar-bg":{bg:"$primary"}},secondary:{".nextui-avatar-bg":{bg:"$secondary"}},success:{".nextui-avatar-bg":{bg:"$success"}},warning:{".nextui-avatar-bg":{bg:"$warning"}},error:{".nextui-avatar-bg":{bg:"$error"}},gradient:{".nextui-avatar-bg":{bg:"$gradient"}}},textColor:{default:{".nextui-avatar-text":{color:"$text"}},white:{".nextui-avatar-text":{color:"$white"}},primary:{".nextui-avatar-text":{color:"$primary"}},secondary:{".nextui-avatar-text":{color:"$secondary"}},success:{".nextui-avatar-text":{color:"$success"}},warning:{".nextui-avatar-text":{color:"$warning"}},error:{".nextui-avatar-text":{color:"$error"}}},size:{xs:{$$avatarXs:"$space$9",sizeMin:"$$avatarXs",".nextui-avatar-text":{fontSize:"$xs"}},sm:{$$avatarSm:"$space$11",sizeMin:"$$avatarSm",".nextui-avatar-text":{fontSize:"$base"}},md:{$$avatarMd:"$space$14",sizeMin:"$$avatarMd",".nextui-avatar-text":{fontSize:"$sm"}},lg:{$$avatarLg:"$space$16",sizeMin:"$$avatarLg",".nextui-avatar-text":{fontSize:"$sm"}},xl:{$$avatarXl:"$space$18",sizeMin:"$$avatarXl",".nextui-avatar-text":{fontSize:"$md"}}},borderWeight:{light:{".nextui-avatar-img":{borderWidth:"$light"}},normal:{".nextui-avatar-img":{borderWidth:"$normal"}},bold:{".nextui-avatar-img":{borderWidth:"$normal"}},extrabold:{".nextui-avatar-img":{borderWidth:"$normal"}},black:{".nextui-avatar-img":{borderWidth:"$normal"}}},bordered:{true:{"&:hover:not(.only-text-avatar) .nextui-avatar-bg":{opacity:"0.6"},".nextui-avatar-img":{borderStyle:"solid",borderColor:"$background"}}},stacked:{true:{ml:"-$5"}},pointer:{true:{cursor:"pointer"}},rounded:{true:{borderRadius:"$rounded",".nextui-avatar-img":{borderRadius:"$rounded"}}},squared:{true:{borderRadius:"$squared",".nextui-avatar-img":{borderRadius:"$squared"}}},zoomed:{true:{"&:hover .nextui-avatar-img":{transform:"scale(1.125)"}}}},compoundVariants:[{bordered:!0,borderWeight:"light",css:{padding:"calc($1/2)"}},{bordered:!0,borderWeight:"normal",css:{padding:"$1"}},{bordered:!0,borderWeight:"bold",css:{padding:"calc($2/1.5)"}},{bordered:!0,borderWeight:"extrabold",css:{padding:"$2"}},{bordered:!0,borderWeight:"black",css:{padding:"calc($3/1.5)"}},{rounded:!0,squared:!0,css:{borderRadius:"$squared",".nextui-avatar-img":{borderRadius:"$squared"}}}],defaultVariants:{size:"md",rounded:!0,color:"default",textColor:"default",borderWeight:"normal"}}),c=r(1309);const u=e=>(null==e?void 0:e.length)<=4?e:null==e?void 0:e.slice(0,3),d=({src:e,text:t,icon:r,alt:n,className:o,...s})=>{const d=!e,[f,p]=(0,i.useState)(!1),b=(0,i.useRef)(null);(0,i.useEffect)((()=>{var e;(null==b||null==(e=b.current)?void 0:e.complete)&&p(!0)}),[]);const x=(0,i.useMemo)((()=>!f&&e?"loading":"ready"),[e,f]);return(0,a.jsxs)(l,{className:(0,c.Z)({"only-text-avatar":d},o),"data-state":x,...s,children:[(0,a.jsx)("span",{className:"nextui-avatar-bg"}),!d&&(0,a.jsx)("img",{ref:b,className:(0,c.Z)("nextui-avatar-img",`nextui-avatar--${x}`,{"nextui-avatar-ready":f}),src:e,alt:n,"data-state":x,onLoad:()=>p(!0)}),d&&!r&&t&&(0,a.jsx)("span",{className:"nextui-avatar-text",children:u(t)}),r&&(0,a.jsx)("span",{className:"nextui-avatar-icon",children:r})]})};d.toString=()=>".nextui-avatar";var f=d;const p=(0,s.zo)("span",{fontSize:"$xs",display:"inline-flex",alignItems:"center",marginLeft:"$3",color:"$text"});var b=(0,s.zo)("div",{dflex:"center",height:"auto",width:"max-content","@motion":{transition:"none"},[`& ${l}`]:{marginLeft:"-$space$5",transition:"$default",".only-text-avatar":{boxShadow:"$xs"}},".only-text-avatar":{boxShadow:"$xs"},variants:{animated:{true:{[`& ${l}:hover`]:{transform:"translate(-$space$5)"}}}},defaultVariants:{animated:!0}});const x=({count:e,children:t,...r})=>(0,a.jsxs)(b,{...r,children:[t,e&&(0,a.jsxs)(p,{className:"nextui-avatar-group-count",children:["+",e]})]});x.toString=()=>".nextui-avatar-group";var m=x;f.Group=m;var h=f,g=r(9693),v=r(6979),y=r(1614),j=r(2553),$=r(8947),w=r(5697),_=r.n(w);function O(e,t){var r=Object.keys(e);if(Object.getOwnPropertySymbols){var a=Object.getOwnPropertySymbols(e);t&&(a=a.filter((function(t){return Object.getOwnPropertyDescriptor(e,t).enumerable}))),r.push.apply(r,a)}return r}function I(e){for(var t=1;t<arguments.length;t++){var r=null!=arguments[t]?arguments[t]:{};t%2?O(Object(r),!0).forEach((function(t){k(e,t,r[t])})):Object.getOwnPropertyDescriptors?Object.defineProperties(e,Object.getOwnPropertyDescriptors(r)):O(Object(r)).forEach((function(t){Object.defineProperty(e,t,Object.getOwnPropertyDescriptor(r,t))}))}return e}function S(e){return S="function"==typeof Symbol&&"symbol"==typeof Symbol.iterator?function(e){return typeof e}:function(e){return e&&"function"==typeof Symbol&&e.constructor===Symbol&&e!==Symbol.prototype?"symbol":typeof e},S(e)}function k(e,t,r){return t in e?Object.defineProperty(e,t,{value:r,enumerable:!0,configurable:!0,writable:!0}):e[t]=r,e}function N(e,t){if(null==e)return{};var r,a,n=function(e,t){if(null==e)return{};var r,a,n={},o=Object.keys(e);for(a=0;a<o.length;a++)r=o[a],t.indexOf(r)>=0||(n[r]=e[r]);return n}(e,t);if(Object.getOwnPropertySymbols){var o=Object.getOwnPropertySymbols(e);for(a=0;a<o.length;a++)r=o[a],t.indexOf(r)>=0||Object.prototype.propertyIsEnumerable.call(e,r)&&(n[r]=e[r])}return n}function Z(e){return function(e){if(Array.isArray(e))return z(e)}(e)||function(e){if("undefined"!==typeof Symbol&&null!=e[Symbol.iterator]||null!=e["@@iterator"])return Array.from(e)}(e)||function(e,t){if(!e)return;if("string"===typeof e)return z(e,t);var r=Object.prototype.toString.call(e).slice(8,-1);"Object"===r&&e.constructor&&(r=e.constructor.name);if("Map"===r||"Set"===r)return Array.from(e);if("Arguments"===r||/^(?:Ui|I)nt(?:8|16|32)(?:Clamped)?Array$/.test(r))return z(e,t)}(e)||function(){throw new TypeError("Invalid attempt to spread non-iterable instance.\nIn order to be iterable, non-array objects must have a [Symbol.iterator]() method.")}()}function z(e,t){(null==t||t>e.length)&&(t=e.length);for(var r=0,a=new Array(t);r<t;r++)a[r]=e[r];return a}function A(e){return t=e,(t-=0)===t?e:(e=e.replace(/[\-_\s]+(.)?/g,(function(e,t){return t?t.toUpperCase():""}))).substr(0,1).toLowerCase()+e.substr(1);var t}var P=["style"];function C(e){return e.split(";").map((function(e){return e.trim()})).filter((function(e){return e})).reduce((function(e,t){var r,a=t.indexOf(":"),n=A(t.slice(0,a)),o=t.slice(a+1).trim();return n.startsWith("webkit")?e[(r=n,r.charAt(0).toUpperCase()+r.slice(1))]=o:e[n]=o,e}),{})}var D=!1;try{D=!0}catch(x){}function T(e){return e&&"object"===S(e)&&e.prefix&&e.iconName&&e.icon?e:$.Qc.icon?$.Qc.icon(e):null===e?null:e&&"object"===S(e)&&e.prefix&&e.iconName?e:Array.isArray(e)&&2===e.length?{prefix:e[0],iconName:e[1]}:"string"===typeof e?{prefix:"fas",iconName:e}:void 0}function W(e,t){return Array.isArray(t)&&t.length>0||!Array.isArray(t)&&t?k({},e,t):{}}var E=["forwardedRef"];function R(e){var t=e.forwardedRef,r=N(e,E),a=r.icon,n=r.mask,o=r.symbol,i=r.className,s=r.title,l=r.titleId,c=r.maskId,u=T(a),d=W("classes",[].concat(Z(function(e){var t,r=e.beat,a=e.fade,n=e.beatFade,o=e.bounce,i=e.shake,s=e.flash,l=e.spin,c=e.spinPulse,u=e.spinReverse,d=e.pulse,f=e.fixedWidth,p=e.inverse,b=e.border,x=e.listItem,m=e.flip,h=e.size,g=e.rotation,v=e.pull,y=(k(t={"fa-beat":r,"fa-fade":a,"fa-beat-fade":n,"fa-bounce":o,"fa-shake":i,"fa-flash":s,"fa-spin":l,"fa-spin-reverse":u,"fa-spin-pulse":c,"fa-pulse":d,"fa-fw":f,"fa-inverse":p,"fa-border":b,"fa-li":x,"fa-flip-horizontal":"horizontal"===m||"both"===m,"fa-flip-vertical":"vertical"===m||"both"===m},"fa-".concat(h),"undefined"!==typeof h&&null!==h),k(t,"fa-rotate-".concat(g),"undefined"!==typeof g&&null!==g&&0!==g),k(t,"fa-pull-".concat(v),"undefined"!==typeof v&&null!==v),k(t,"fa-swap-opacity",e.swapOpacity),t);return Object.keys(y).map((function(e){return y[e]?e:null})).filter((function(e){return e}))}(r)),Z(i.split(" ")))),f=W("transform","string"===typeof r.transform?$.Qc.transform(r.transform):r.transform),p=W("mask",T(n)),b=(0,$.qv)(u,I(I(I(I({},d),f),p),{},{symbol:o,title:s,titleId:l,maskId:c}));if(!b)return function(){var e;!D&&console&&"function"===typeof console.error&&(e=console).error.apply(e,arguments)}("Could not find icon",u),null;var x=b.abstract,m={ref:t};return Object.keys(r).forEach((function(e){R.defaultProps.hasOwnProperty(e)||(m[e]=r[e])})),M(x[0],m)}R.displayName="FontAwesomeIcon",R.propTypes={beat:_().bool,border:_().bool,bounce:_().bool,className:_().string,fade:_().bool,flash:_().bool,mask:_().oneOfType([_().object,_().array,_().string]),maskId:_().string,fixedWidth:_().bool,inverse:_().bool,flip:_().oneOf(["horizontal","vertical","both"]),icon:_().oneOfType([_().object,_().array,_().string]),listItem:_().bool,pull:_().oneOf(["right","left"]),pulse:_().bool,rotation:_().oneOf([0,90,180,270]),shake:_().bool,size:_().oneOf(["2xs","xs","sm","lg","xl","2xl","1x","2x","3x","4x","5x","6x","7x","8x","9x","10x"]),spin:_().bool,spinPulse:_().bool,spinReverse:_().bool,symbol:_().oneOfType([_().bool,_().string]),title:_().string,titleId:_().string,transform:_().oneOfType([_().string,_().object]),swapOpacity:_().bool},R.defaultProps={border:!1,className:"",mask:null,maskId:null,fixedWidth:!1,inverse:!1,flip:null,icon:null,listItem:!1,pull:null,pulse:!1,rotation:null,size:null,spin:!1,beat:!1,fade:!1,beatFade:!1,bounce:!1,shake:!1,symbol:!1,title:"",titleId:null,transform:null,swapOpacity:!1};var M=function e(t,r){var a=arguments.length>2&&void 0!==arguments[2]?arguments[2]:{};if("string"===typeof r)return r;var n=(r.children||[]).map((function(r){return e(t,r)})),o=Object.keys(r.attributes||{}).reduce((function(e,t){var a=r.attributes[t];switch(t){case"class":e.attrs.className=a,delete r.attributes.class;break;case"style":e.attrs.style=C(a);break;default:0===t.indexOf("aria-")||0===t.indexOf("data-")?e.attrs[t.toLowerCase()]=a:e.attrs[A(t)]=a}return e}),{attrs:{}}),i=a.style,s=void 0===i?{}:i,l=N(a,P);return o.attrs.style=I(I({},o.attrs.style),s),t.apply(void 0,[r.tag,I(I({},o.attrs),l)].concat(Z(n)))}.bind(null,i.createElement),H=r(1417),L=r(3020),q=r.n(L),F=function(){return(0,a.jsxs)(n.Z,{className:q().aboutMe,display:"flex",justify:"center",direction:"column",alignItems:"center",alignContent:"center",children:[(0,a.jsx)(h,{className:q().avatar,src:"images/AM_Pic_3.png",css:{size:"$72"},zoomed:!0}),(0,a.jsx)(g.Z,{}),(0,a.jsx)(v.Z,{h1:!0,weight:"extrabold",children:"Aaron Mechanic"}),(0,a.jsxs)(v.Z,{h3:!0,children:["Software Developer II - ",(0,a.jsx)("a",{className:q().springLink,href:"https://springhealth.com/",target:"_blank",rel:"noreferrer noopener",children:"Spring Health"})]}),(0,a.jsx)(v.Z,{h6:!0,children:"B.S.Computer Science - CUNY Queens college 2021"}),(0,a.jsx)(g.Z,{}),(0,a.jsx)(n.Z,{className:q().bio,children:(0,a.jsxs)(y.ZP,{bordered:!0,children:[(0,a.jsx)(v.Z,{className:q().bioText,children:"Hi! I'm Aaron, a software engineer from New York City and I've been a computer fanatic my whole life. I fell in love with computers ever since my first, an iMac G3. I decided to become a programmer because I love identifying and solving complex problems, and I love learning new things. Every week 300,000 new Javascript frameworks are released and I learn them all. Haha just kidding."}),(0,a.jsx)(g.Z,{}),(0,a.jsx)(v.Z,{className:q().bioText,children:"In 2021, at the height of the pandemic, I graduated from CUNY Queens\n                    college with a Bachelors in Computer Science. In June 2022 I started as a Software Development\n                    Intern for Spring Health. In September I was hired full time as a Software Developer I. Currently, I\n                    am employed by Spring Health as a Software Developer 2."})]})}),(0,a.jsx)(g.Z,{y:1.5}),(0,a.jsxs)(n.Z,{className:q().social_links,display:"flex",justify:"space-evenly",alignItems:"center",alignContent:"center",children:[(0,a.jsx)("a",{href:"https://github.com/aer35",target:"_blank",rel:"noreferrer noopener",children:(0,a.jsx)(R,{size:"4x",icon:H.zhw})}),(0,a.jsx)("a",{href:"https://www.linkedin.com/in/aaronmechanic/",target:"_blank",rel:"noreferrer noopener",children:(0,a.jsx)(R,{size:"4x",icon:H.D9H})}),(0,a.jsx)("a",{href:"https://angel.co/u/aaron-mechanic",target:"_blank",rel:"noreferrer noopener",children:(0,a.jsx)(R,{size:"4x",icon:H.dFc})})]}),(0,a.jsx)(g.Z,{y:1.5}),(0,a.jsx)(n.Z,{className:q().bio,children:(0,a.jsx)(y.ZP,{bordered:!0,children:(0,a.jsxs)(n.Z,{className:q().history,children:[(0,a.jsx)(v.Z,{h3:!0,children:"Work History"}),(0,a.jsxs)("ul",{children:[(0,a.jsx)("li",{children:(0,a.jsxs)(n.Z,{className:q().jobDetails,children:[(0,a.jsx)(v.Z,{h4:!0,children:"Spring Health"}),(0,a.jsxs)("ul",{children:[(0,a.jsx)("li",{children:(0,a.jsxs)(j.ZP,{className:q().jobItem,children:[(0,a.jsx)(v.Z,{h5:!0,children:"Software Developer 2"}),(0,a.jsx)(n.Z,{className:q().jobDate,children:(0,a.jsx)(v.Z,{h5:!0,children:"June 2023 - Current"})})]})}),(0,a.jsx)("li",{children:(0,a.jsxs)(j.ZP,{className:q().jobItem,children:[(0,a.jsx)(v.Z,{h6:!0,children:"Software Developer 1"}),(0,a.jsx)(n.Z,{className:q().jobDate,children:(0,a.jsx)(v.Z,{h6:!0,children:"August 2022 - June 2023"})})]})}),(0,a.jsx)("li",{children:(0,a.jsxs)(j.ZP,{className:q().jobItem,children:[(0,a.jsx)(v.Z,{h6:!0,children:"Software Development Intern"}),(0,a.jsx)(n.Z,{className:q().jobDate,children:(0,a.jsx)(v.Z,{h6:!0,children:"June 2022 - August 2022"})})]})})]})]})}),(0,a.jsx)("li",{children:(0,a.jsxs)(j.ZP,{className:q().jobItem,children:[(0,a.jsx)(n.Z,{className:q().jobDetails,children:(0,a.jsx)(v.Z,{children:"Hello World"})}),(0,a.jsx)(n.Z,{className:q().jobDate,children:(0,a.jsx)(v.Z,{children:"January 1900"})})]})})]})]})})})]})},J=function(){return(0,a.jsxs)(a.Fragment,{children:[(0,a.jsx)(o.default,{children:(0,a.jsx)("title",{children:"Home"})}),(0,a.jsx)(n.Z,{style:{minHeight:"80vh"},children:(0,a.jsx)(F,{})})]})}},3020:function(e){e.exports={aboutMe:"About_aboutMe__UF5s6",avatar:"About_avatar__IP4nt",springLink:"About_springLink__S9ivA",social_links:"About_social_links__KDXAX",bio:"About_bio__DA07_",bioText:"About_bioText__eYIA7",history:"About_history__JmA39",jobItem:"About_jobItem__Zwave",jobDate:"About_jobDate__2S1JV"}},2703:function(e,t,r){"use strict";var a=r(414);function n(){}function o(){}o.resetWarningCache=n,e.exports=function(){function e(e,t,r,n,o,i){if(i!==a){var s=new Error("Calling PropTypes validators directly is not supported by the `prop-types` package. Use PropTypes.checkPropTypes() to call them. Read more at http://fb.me/use-check-prop-types");throw s.name="Invariant Violation",s}}function t(){return e}e.isRequired=e;var r={array:e,bigint:e,bool:e,func:e,number:e,object:e,string:e,symbol:e,any:e,arrayOf:t,element:e,elementType:e,instanceOf:t,node:e,objectOf:t,oneOf:t,oneOfType:t,shape:t,exact:t,checkPropTypes:o,resetWarningCache:n};return r.PropTypes=r,r}},5697:function(e,t,r){e.exports=r(2703)()},414:function(e){"use strict";e.exports="SECRET_DO_NOT_PASS_THIS_OR_YOU_WILL_BE_FIRED"}},function(e){e.O(0,[112,374,774,888,179],(function(){return t=5301,e(e.s=t);var t}));var t=e.O();_N_E=t}]);
