+++ pages/index.tsx
@@ -1,7 +1,7 @@
 import { Container } from "@nextui-org/react";
 import type { NextPage } from "next";
 import Head from "next/head";
-import AboutMe from "../components/AboutMe";
+import AboutMe from "../components/about_me/AboutMe";
 
 const Home: NextPage = () => {
   return (
